Fixed Title using old window title
Fixed occasional error that would put the previous page's title in the video title - even if it was a channel name. Optimized title fetching and made it update the title every iteration to be sure. Also reverted page change detection to title, since it does work with autoplay and that makes the script realize a page change later, which solves some problems of youtube not loading content fast enough.